retornando erro aos recursos não encontrados

+++ app/Http/Controllers/MarcaController.php
@@ -8,24 +8,24 @@ use Illuminate\Http\Request;
 class MarcaController extends Controller
 {
     /**
+     * Injecting the Model instance as contructor method .
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        $marcas = Marca::all();
+        $marcas['marcas'] = $this->marca->all();
         return $marcas;
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
     }
 
     /**
@@ -36,53 +36,54 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
+        $marca['marca'] = $this->marca->create($request->all());
         return $marca;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
+        $marca['marca'] = $this->marca->find($id);
+        if ($marca['marca'] === null) {
+            return ["error" => "Not Found"];
+        }
         return $marca;
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Marca  $marca
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Marca $marca)
-    {
-        //
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
-        $marca->update($request->all());
+        $marca['marca'] = $this->marca->find($id);
+        if ($marca['marca'] === null) {
+            return ["error" => "Not Found"];
+        }
+        $marca['marca']->update($request->all());
         return $marca;
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+        if ($marca['marca'] === null) {
+            return ["error" => "Not Found"];
+        }
         $marca->delete();
         $marca['removido'] = true;
         return $marca;
